move header into app folder
