+++ src/renderer/app/header/name_change_dialog.tsx
