+++ src/renderer/app/header/start_game_dialog.tsx
