+++ src/renderer/app/header/user_info/user_info.stories.tsx
@@ -1,6 +1,6 @@
 import { generateDisplayPicture } from "@/lib/displayPicture";
 
-import { UserInfo } from "./UserInfo";
+import { UserInfo } from "./user_info";
 
 // type UserInfoProps = React.ComponentProps<typeof UserInfo>;
