+++ src/renderer/app/header/user_info/user_info.tsx
